scope_test_top.vhd, usbadc-support module: hold of writing to buffer
Initially neither the (SDRAM) sample buffer nor the ADC clock are availeble.
The SDRAM needs to be initialized and the PLL programmed.

Added 'adcRst' signal to prevent data from being acquired until adcRst
is released (from top-level, when PLL is stable).

The SDRAM buffer cannot be written until the SDRAM is ready.